0010: Fix eslint error & small refactors
