+++ src/hooks/use-auth.ts
@@ -1,0 +1,8 @@
+import { AuthContext } from "@/providers/AuthProvider";
+import { useContext } from "react";
+
+function useAuth() {
+  return useContext(AuthContext);
+}
+
+export { useAuth };
