+++ src/hooks/use-clubs.ts
@@ -1,0 +1,8 @@
+import { ClubContext } from "@/providers/ClubProvider";
+import { useContext } from "react";
+
+function useClubs() {
+  return useContext(ClubContext);
+}
+
+export { useClubs };
